# Conflicts:
#	package/cloudshell/cp/vcenter/commands/command_orchestrator.py

+++ package/cloudshell/cp/vcenter/commands/destroy_vm.py
@@ -31,8 +31,14 @@ class DestroyVirtualMachineCommand(object):
                                            logger=logger)
         # find vm
         vm = self.pv_service.find_by_uuid(si, vm_uuid)
-        # destroy vm
-        result = self.pv_service.destroy_vm(vm=vm, logger=logger)
+
+        if vm is not None:
+            # destroy vm
+            result = self.pv_service.destroy_vm(vm=vm, logger=logger)
+        else:
+            logger.info("Could not find the VM {0},will remove the resource.".format(vm_name))
+            result = True
+
         # delete resources
         self.resource_remover.remove_resource(session=session, resource_full_name=vm_name)
         return result
